requirements.txt
numpy
scipy
scikit-learn
matplotlib
joblib
soundfile
torch
transformers
modelscope
umap-learn
